move resolution logic to its own file
